+++ www/modules/email/MessagesGrid.js
@@ -166,10 +166,6 @@ GO.email.MessagesGrid = function(config){
 							}
 
 							if(this.searchField && this.searchField.getValue()) {
-								// GO.email.messagesGrid.store.baseParams['search'] = this.searchField.getValue();
-								// this.searchField.hasSearch = true;
-
-								// GO.email.messagesGrid.store.reload();
 								this.searchField.search();
 							}
 						}
@@ -224,12 +220,11 @@ GO.email.MessagesGrid = function(config){
 	this.showUnreadButton = new Ext.Button({
 		iconCls: 'ic-markunread',
 		enableToggle:true,
-		toggleHandler:this.toggleUnread,
+		listeners: {toggle: this.toggleUnread,scope:this},
 		pressed:false,
 		tooltip: t("Show unread", "email")
 	});
 	this.showFlaggedButton = new Ext.menu.CheckItem({
-		//iconCls: 'ic-flag',
 		enableToggle:true,
 		listeners: {checkchange: this.toggleFlagged,scope:this},
 		pressed:false,
@@ -264,8 +259,6 @@ GO.email.MessagesGrid = function(config){
 	GO.email.MessagesGrid.superclass.constructor.call(this, config);
 
 	var me = this;
-
-
 
 	if(!config.hideSearch) {
 		this.getTopToolbar().enableOverflow = true;
@@ -311,7 +304,6 @@ GO.email.MessagesGrid = function(config){
 			'->',
 			this.showUnreadButton,
 			this.searchField = new go.toolbar.SearchButton({
-				//store: config.store,
 				paramName: 'search',
 				hidden: config.hideSearch,
 				tools: [
@@ -344,7 +336,6 @@ GO.email.MessagesGrid = function(config){
 		);
 	}
 
-
 	var origRefreshHandler = this.getBottomToolbar().refresh.handler;
 
 	this.getBottomToolbar().refresh.handler=function(){
@@ -355,7 +346,7 @@ GO.email.MessagesGrid = function(config){
 
 	//stop/start drag and drop when store loads when account is readOnly
 	this.store.on('load', function(store, records, options) {
-		console.log(store.baseParams);
+		// console.log(store.baseParams);
 		if(this.getView().dragZone){
 			if(store.reader.jsonData.permission_level <= GO.permissionLevels.read) {
 				this.getView().dragZone.lock();
@@ -364,8 +355,6 @@ GO.email.MessagesGrid = function(config){
 			}
 		}
 	}, this);
-
-
 
 }
 
@@ -405,50 +394,20 @@ Ext.extend(GO.email.MessagesGrid, go.grid.GridPanel,{
 	},
 	toggleUnread : function(item, pressed)
 	{
-		this.setIconClass(pressed ? 'ic-email' : 'ic-mark-as-unread');
-		this.setTooltip(pressed ? t("Show all", "email") : t("Show unread", "email"));
-		GO.email.messagesGrid.store.baseParams['unread']=pressed ? 1 : 0;
-
-		GO.email.messagesGrid.store.load();
+		item.setIconClass(pressed ? 'ic-email' : 'ic-mark-as-unread');
+		item.setTooltip(pressed ? t("Show all", "email") : t("Show unread", "email"));
+		// GO.email.messagesGrid.store.baseParams['unread']=pressed ? 1 : 0;
+		// GO.email.messagesGrid.store.load();
+		this.store.baseParams['unread'] = pressed ? 1 : 0;
+		this.store.load();
 	},
 	toggleFlagged : function(item, pressed)
 	{
-		GO.email.messagesGrid.store.baseParams['flagged']=pressed ? 1 : 0;
-		
-		GO.email.messagesGrid.store.load();
+		// GO.email.messagesGrid.store.baseParams['flagged']=pressed ? 1 : 0;
+		// GO.email.messagesGrid.store.load();
+		this.store.baseParams['flagged']=pressed ? 1 : 0;
+		this.store.load();
 	},
-
-	/* @deprecated
-	renderNorthMessageRow : function(value, metaData, record){
-
-		if( this.isSpoofed(record)) {
-			metaData.css = 'danger';
-
-			value += " &lt;" + record.data.sender + "&gt;";
-		}
-
-		if(record.data['seen']=='0')
-			return String.format('<div id="sbj_'+record.data['uid']+'" '+this.createQtipTemplate(record)+' class="ml-unseen-mail">{0}</div>', value);
-		else
-			return String.format('<div id="sbj_'+record.data['uid']+'" '+this.createQtipTemplate(record)+' class="ml-seen-mail">{0}</div>', value);
-	},
-	*/
-	/* @deprecated
-	renderMessageSmallRes : function(value, metaData, record){
-
-		if( this.isSpoofed(record)) {
-			metaData.css = 'danger';
-			value += " &lt;" + record.data.sender + "&gt;";
-		}
-
-		if(record.data['seen']=='0')
-		{
-			return String.format('<div id="sbj_'+record.data['uid']+'" '+this.createQtipTemplate(record)+' class="ml-unseen-from">{0}</div><div class="ml-unseen-subject">{1}</div>', value, record.data['subject']);
-		}else
-		{
-			return String.format('<div id="sbj_'+record.data['uid']+'" '+this.createQtipTemplate(record)+' class="ml-seen-from">{0}</div><div class="ml-seen-subject">{1}</div>', value, record.data['subject']);
-		}
-	},*/
 
 	createQtipTemplate: function(record){
 		var qtipTemplate = '';
@@ -486,8 +445,8 @@ Ext.extend(GO.email.MessagesGrid, go.grid.GridPanel,{
 	},
 
 	renderIcon : function(src, p, record){
-		var icons = [];
-		var unseen = '';
+		let icons = [];
+		let unseen = '';
 		if(record.data.answered) {
 			icons.push('reply');
 		}
@@ -495,7 +454,7 @@ Ext.extend(GO.email.MessagesGrid, go.grid.GridPanel,{
 			icons.push('forward');
 		}
 		if(!record.data.seen) {
-			var unseen = '<div class="ml-unseen-dot"></div>';
+			unseen = '<div class="ml-unseen-dot"></div>';
 		}
 		if(record.data['has_attachments']=='1') {
 			icons.push('attachment');
@@ -515,25 +474,5 @@ Ext.extend(GO.email.MessagesGrid, go.grid.GridPanel,{
 			return '<i class="icon '+(i!=='flag'?'c-secondary':'red')+'">' + i + '</i>';
 		}).join("");
 		
-	},
-
-	/* @deprecated
-	renderFlagged : function(value, p, record){
-
-		var str = '';
-
-		if(record.data['flagged']==1)
-		{
-			//str += '<img src=\"' + GOimages['flag'] +' \" style="display:block" />';
-			str += '<div class="go-icon btn-flag"></div>';
-		}
-		if(record.data['attachments'])
-		{
-			str += '<div class="go-icon btn-attach"></div>';
-		//str += '<img src=\"' + GOimages['attach'] +' \" style="display:block" />';
-		}
-		return str;
-
 	}
-	 */
 });