+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Email: fix message filters for read/unread and flagged
+
 22-04-2024: 6.7.76
 - Tickets: note messages will no longer be sent to the CC-address of the ticket.
 - Email: Save and Send button will be disabled while attachment upload is in progress.